requirements.txt
openai
chromadb
flask
requests
python-dotenv
sentence-transformers
flask-cors